requirements.txt
asgiref==3.5.2
attrs==22.1.0
beautifulsoup4==4.11.1
black==22.6.0
bs4==0.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
coreapi==2.3.3
coreschema==0.0.4
coverage==6.5.0
cryptography==38.0.1
defusedxml==0.7.1
Django==4.1
django-autoslug==1.9.8
django-countries==7.3.2
django-environ==0.9.0
django-filter==22.1
django-phonenumber-field==6.4.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
factory-boy==3.2.1
Faker==15.0.0
flake8==5.0.4
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.7.0
mypy-extensions==0.4.3
oauthlib==3.2.1
packaging==21.3
pathspec==0.9.0
phonenumbers==8.12.54
Pillow==9.2.0
pip-search==0.0.12
platformdirs==2.5.2
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
Pygments==2.13.0
PyJWT==2.1.0
pyparsing==3.0.9
pytest==7.1.3
pytest-cov==4.0.0
pytest-django==4.5.2
pytest-factoryboy==2.5.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.5.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.12
